requirements.txt
python-can
RPi.GPIO
Flask
pytz
cantools
paho-mqtt
protobuf